+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - change `GooseTask` signature to take a mutable reference of `GooseUser` (API change)
  - remove `Clone` trait from `GooseUser` and `GooseAttack`
  - update `GooseTaskSet::set_wait_time()` to accept `std::time::Duration` instead of `usize` allowing more granularity (API change)
+ - use request name when displaying errors to avoid having a large volume of distinct error for the same endpoint when using path parameters
 
 ## 0.13.3 August 25, 2021
  - document GooseConfiguration fields that were only documented as gumpdrop parameters (in order to generate new lines in the help output) so now they're also documented in the code